# api/openapi.yaml
openapi: 3.1.0
info:
  title: QPay Merchant API
  version: 1.0.0
  description: |
    Initial public contract for merchant Pay-In, wallet, beneficiary, payout,
    refund, and transaction-tracking operations. Monetary amounts are minor units.
servers:
  - url: https://api.qpay.example.com/api/v1
    description: Production placeholder
tags:
  - name: Authentication
  - name: Wallets
  - name: Payments
  - name: Beneficiaries
  - name: Payouts
  - name: Transactions
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate a user
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Token pair issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Rotate a refresh token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  minLength: 32
      responses:
        '200':
          description: Rotated token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /wallets/{walletId}:
    get:
      tags: [Wallets]
      summary: Get a wallet balance projection
      operationId: getWallet
      parameters:
        - $ref: '#/components/parameters/WalletId'
      responses:
        '200':
          description: Wallet found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        '404':
          $ref: '#/components/responses/NotFound'
  /wallets/{walletId}/transactions:
    get:
      tags: [Wallets]
      summary: List wallet ledger transactions
      operationId: listWalletTransactions
      parameters:
        - $ref: '#/components/parameters/WalletId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Transaction page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionPage'
  /payments:
    post:
      tags: [Payments]
      summary: Create a Pay-In
      operationId: createPayment
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: Payment created
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
  /payments/{paymentId}:
    get:
      tags: [Payments]
      summary: Get payment state
      operationId: getPayment
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      responses:
        '200':
          description: Payment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '404':
          $ref: '#/components/responses/NotFound'
  /payments/{paymentId}/refunds:
    post:
      tags: [Payments]
      summary: Request a full or partial refund
      operationId: createRefund
      parameters:
        - $ref: '#/components/parameters/PaymentId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRefundRequest'
      responses:
        '202':
          description: Refund accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Refund'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
  /utility-payments:
    post:
      tags: [Payments]
      summary: Pay a utility bill from a wallet
      operationId: createUtilityPayment
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUtilityPaymentRequest' }
      responses:
        '202':
          description: Utility payment accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UtilityPayment' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /beneficiaries:
    post:
      tags: [Beneficiaries]
      summary: Register a payout beneficiary
      operationId: createBeneficiary
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBeneficiaryRequest'
      responses:
        '201':
          description: Beneficiary registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Beneficiary'
        '409':
          $ref: '#/components/responses/Conflict'
  /payouts:
    post:
      tags: [Payouts]
      summary: Create a payout
      operationId: createPayout
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePayoutRequest'
      responses:
        '202':
          description: Payout accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payout'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
  /payouts/{payoutId}:
    get:
      tags: [Payouts]
      summary: Get payout state
      operationId: getPayout
      parameters:
        - $ref: '#/components/parameters/PayoutId'
      responses:
        '200':
          description: Payout found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payout'
        '404':
          $ref: '#/components/responses/NotFound'
  /transactions/{reference}:
    get:
      tags: [Transactions]
      summary: Track a transaction by QPay or merchant reference
      operationId: trackTransaction
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 100
      responses:
        '200':
          description: Transaction found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Unique key scoped to merchant and operation.
      schema:
        type: string
        minLength: 8
        maxLength: 190
    WalletId:
      name: walletId
      in: path
      required: true
      schema: { type: string, format: uuid }
    PaymentId:
      name: paymentId
      in: path
      required: true
      schema: { type: string, format: uuid }
    PayoutId:
      name: payoutId
      in: path
      required: true
      schema: { type: string, format: uuid }
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: Authentication failed
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: Idempotency conflict or invalid state transition
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unprocessable:
      description: Valid request that violates a business rule
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
    Money:
      type: object
      additionalProperties: false
      required: [amountMinor, currency]
      properties:
        amountMinor: { type: integer, format: int64 }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: { type: string, minLength: 3, maxLength: 190 }
        password: { type: string, minLength: 8, maxLength: 200 }
    TokenResponse:
      type: object
      additionalProperties: false
      required: [accessToken, refreshToken, tokenType, expiresIn]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        tokenType: { type: string, const: Bearer }
        expiresIn: { type: integer, description: Access-token lifetime in seconds. }
    Wallet:
      type: object
      additionalProperties: false
      required: [id, ownerId, status, available, held, version, updatedAt]
      properties:
        id: { type: string, format: uuid }
        ownerId: { type: string, format: uuid }
        status: { type: string, enum: [ACTIVE, FROZEN, CLOSED] }
        available: { $ref: '#/components/schemas/Money' }
        held: { $ref: '#/components/schemas/Money' }
        version: { type: integer, format: int64 }
        updatedAt: { type: string, format: date-time }
    CreatePaymentRequest:
      type: object
      additionalProperties: false
      required: [walletId, merchantReference, amount, paymentMethod, returnUrl]
      properties:
        walletId: { type: string, format: uuid }
        merchantReference: { type: string, minLength: 1, maxLength: 100 }
        amount: { $ref: '#/components/schemas/Money' }
        paymentMethod:
          type: object
          additionalProperties: false
          required: [type]
          properties:
            type: { type: string, enum: [CARD, BANK_TRANSFER, UPI, WALLET] }
        returnUrl: { type: string, format: uri, maxLength: 2048 }
        description: { type: string, maxLength: 500 }
        metadata:
          type: object
          maxProperties: 20
          additionalProperties: { type: string, maxLength: 500 }
    Payment:
      type: object
      additionalProperties: false
      required: [id, merchantReference, amount, fee, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        merchantReference: { type: string }
        amount: { $ref: '#/components/schemas/Money' }
        fee: { $ref: '#/components/schemas/Money' }
        status:
          type: string
          enum: [CREATED, PROCESSING, REQUIRES_ACTION, SUCCEEDED, FAILED, EXPIRED, PARTIALLY_REFUNDED, REFUNDED]
        nextAction:
          oneOf:
            - type: 'null'
            - type: object
              additionalProperties: false
              required: [type, url]
              properties:
                type: { type: string, enum: [REDIRECT] }
                url: { type: string, format: uri }
        providerReference: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateRefundRequest:
      type: object
      additionalProperties: false
      required: [merchantReference, amount]
      properties:
        merchantReference: { type: string, minLength: 1, maxLength: 100 }
        amount: { $ref: '#/components/schemas/Money' }
        reason: { type: string, maxLength: 500 }
    Refund:
      type: object
      additionalProperties: false
      required: [id, paymentId, merchantReference, amount, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        paymentId: { type: string, format: uuid }
        merchantReference: { type: string }
        amount: { $ref: '#/components/schemas/Money' }
        status: { type: string, enum: [PENDING, PROCESSING, SUCCEEDED, FAILED] }
        createdAt: { type: string, format: date-time }
    CreateBeneficiaryRequest:
      type: object
      additionalProperties: false
      required: [merchantReference, name, bankCode, accountNumber]
      properties:
        merchantReference: { type: string, minLength: 1, maxLength: 100 }
        name: { type: string, minLength: 1, maxLength: 190 }
        bankCode: { type: string, minLength: 1, maxLength: 40 }
        accountNumber: { type: string, minLength: 4, maxLength: 64, writeOnly: true }
    Beneficiary:
      type: object
      additionalProperties: false
      required: [id, merchantReference, name, bankCode, accountMask, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        merchantReference: { type: string }
        name: { type: string }
        bankCode: { type: string }
        accountMask: { type: string }
        status: { type: string, enum: [PENDING, ACTIVE, DISABLED] }
        createdAt: { type: string, format: date-time }
    CreateUtilityPaymentRequest:
      type: object
      additionalProperties: false
      required: [walletId, merchantReference, billerCode, consumerReference, amount]
      properties:
        walletId: { type: string, format: uuid }
        merchantReference: { type: string, minLength: 1, maxLength: 100 }
        billerCode: { type: string, minLength: 1, maxLength: 40 }
        consumerReference: { type: string, minLength: 1, maxLength: 190 }
        amount: { $ref: '#/components/schemas/Money' }
    UtilityPayment:
      type: object
      required: [id, merchantReference, billerCode, consumerReference, amountMinor, feeMinor, currency, status]
      properties:
        id: { type: string, format: uuid }
        merchantReference: { type: string }
        billerCode: { type: string }
        consumerReference: { type: string }
        amountMinor: { type: integer, format: int64 }
        feeMinor: { type: integer, format: int64 }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
        status: { type: string, enum: [CREATED, FUNDS_HELD, SUBMITTED, SUCCEEDED, FAILED, FUNDS_RELEASED] }
    CreatePayoutRequest:
      type: object
      additionalProperties: false
      required: [merchantReference, walletId, beneficiaryId, amount]
      properties:
        merchantReference: { type: string, minLength: 1, maxLength: 100 }
        walletId: { type: string, format: uuid }
        beneficiaryId: { type: string, format: uuid }
        amount: { $ref: '#/components/schemas/Money' }
        description: { type: string, maxLength: 500 }
    Payout:
      type: object
      additionalProperties: false
      required: [id, merchantReference, walletId, beneficiaryId, amount, fee, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        merchantReference: { type: string }
        walletId: { type: string, format: uuid }
        beneficiaryId: { type: string, format: uuid }
        amount: { $ref: '#/components/schemas/Money' }
        fee: { $ref: '#/components/schemas/Money' }
        status:
          type: string
          enum: [CREATED, FUNDS_HELD, SUBMITTED, PROCESSING, SUCCEEDED, FAILED, CANCELLED, FUNDS_RELEASED]
        failureCode: { type: [string, 'null'] }
        providerReference: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Transaction:
      type: object
      additionalProperties: false
      required: [id, reference, type, amount, status, occurredAt]
      properties:
        id: { type: string, format: uuid }
        reference: { type: string }
        type: { type: string, enum: [PAYMENT, PAYOUT, REFUND, UTILITY_PAYMENT, ADJUSTMENT] }
        amount: { $ref: '#/components/schemas/Money' }
        status: { type: string }
        description: { type: [string, 'null'] }
        occurredAt: { type: string, format: date-time }
    TransactionPage:
      type: object
      additionalProperties: false
      required: [items, hasMore]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Transaction' }
        nextCursor: { type: [string, 'null'] }
        hasMore: { type: boolean }
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, requestId, timestamp, details]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string }
        timestamp: { type: string, format: date-time }
        details:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              field: { type: string }
              issue: { type: string }